Add content-deprecation.html to include in out-of-date content.
Some content is too old to work on current set of Fedoras.
This notice gives a way to let the community know that those pages
do not contain reliable and current information.
Pages that include the deprecation notice will be removed in 6 months
from adding the notice.

If page will include the notice, a notification email asking
for community action to update the pages.